changes suggested by marius
all the css classes and ids must be prefixed by wppr string, e.g
wppr-btn, wppr-prodrow.

remove the rating type setting. For style default use the round rating
and for style1 use the stars rating. we dont users anymore to control
this.

Add the image checkbox for both of the styles. If the images does not
exists ( either the box is not checked or the image is not available )
remove the palceholder.

Read review should not have blank target attribute, just the buy now.